+++ apps/frontend/src/libs/enums/app-route.enum.ts
@@ -4,6 +4,7 @@ const AppRoute = {
 	INFORMATION: "/information",
 	SIGN_IN: "/sign-in",
 	SIGN_UP: "/sign-up",
+	TOURPAGE: "/tour-page",
 	TOURSPAGE: "/tours-page",
 } as const;
